+++ application/src/electron/main.ts
@@ -49,8 +49,8 @@ export function sendNotification(notification: Notification) {
 function createWindow() {    
     // Create the browser window.
     mainWindow = new BrowserWindow({
-        width: 800,
-        height: 600,
+        width: 1000,
+        height: 700,
         webPreferences: {
             nodeIntegration: true,
             contextIsolation: true,
