+++ application/src/frontend/views/GameInstallView.svelte
@@ -1,9 +1,8 @@
 <script lang="ts">
   import { onMount } from "svelte";
-  import { fetchAddonsWithConfigure, getConfigClientOption, getDownloadPath, safeFetch } from "../utils";
+  import { fetchAddonsWithConfigure, safeFetch, startDownload, type SearchResultWithAddon } from "../utils";
   import type { OGIAddonConfiguration, SearchResult } from "ogi-addon";
   import type { ConfigurationFile } from "ogi-addon/config";
-  import { createNotification, currentDownloads, notifications } from "../store";
 	interface ConfigTemplateAndInfo extends OGIAddonConfiguration {
     configTemplate: ConfigurationFile
   }
@@ -14,9 +13,6 @@
     });
   });
 
-	type SearchResultWithAddon = SearchResult & {
-		addonSource: string
-	}
 
 	let results: SearchResultWithAddon[] = [];
 
@@ -44,261 +40,7 @@
 		}
 	}
 
-	async function startDownload(result: SearchResultWithAddon, event: MouseEvent) {
-		if (event === null) return;
-		if (event.target === null) return;
-		const htmlButton = (event.target as HTMLElement).closest('button')!!;
-		htmlButton.querySelector('[data-dwtext]')!!.textContent = "Downloading...";
-		htmlButton.disabled = true;
-		let downloadType = result.downloadType;
-		if (downloadType === "torrent" || downloadType === 'magnet') {
-			const generalOptions = getConfigClientOption('general') as any;
-			const torrentClient: "webtorrent" | "qbittorrent" | "real-debrid" = (generalOptions ? generalOptions.torrentClient : null) ?? 'webtorrent';
-			if (torrentClient === 'real-debrid') {
-				downloadType = 'real-debrid-' + downloadType;
-			}
-		}
-
-		switch (downloadType) {
-			case 'real-debrid-magnet': {
-				if (!result.downloadURL) {
-					createNotification({
-						id: Math.random().toString(36).substring(7),
-						type: 'error',
-						message: "Addon did not provide a magnet link."
-					});
-					return;
-				}
-				const worked = await window.electronAPI.realdebrid.updateKey();
-				if (!worked) {
-					createNotification({
-						id: Math.random().toString(36).substring(7),
-						type: 'error',
-						message: "Please set your Real-Debrid API key in the settings."
-					});
-					return;
-				}
-				// get the first host
-				const hosts = await window.electronAPI.realdebrid.getHosts();	
-				const localID = Math.floor(Math.random() * 1000000);
-				currentDownloads.update((downloads) => {
-					return [...downloads, { 
-						id: '' + localID, 
-						status: 'rd-downloading', 
-						downloadPath: getDownloadPath() + "\\" + result.name, 
-						downloadSpeed: 0,
-						usedRealDebrid: true,
-						progress: 0,
-						...result 
-					}];
-				});
-				// add magnet link
-				const magnetLink = await window.electronAPI.realdebrid.addMagnet(result.downloadURL, hosts[0]);
-				const isReady = await window.electronAPI.realdebrid.isTorrentReady(magnetLink.id);
-				if (!isReady) {
-					window.electronAPI.realdebrid.selectTorrent(magnetLink.id);
-					await new Promise<void>((resolve) => {
-						const interval = setInterval(async () => {
-							const isReady = await window.electronAPI.realdebrid.isTorrentReady(magnetLink.id);
-							if (isReady) {
-								clearInterval(interval);
-								resolve();
-							}
-						}, 1000);
-					});
-				}
-
-
-				const torrentInfo = await window.electronAPI.realdebrid.getTorrentInfo(magnetLink.id);
-				const download = await window.electronAPI.realdebrid.unrestrictLink(torrentInfo.links[0]);
-
-				if (download === null) {
-					createNotification({
-						id: Math.random().toString(36).substring(7),
-						type: 'error',
-						message: "Failed to unrestrict the link."
-					});
-					return;
-				}
-				const downloadID = await window.electronAPI.ddl.download([ { link: download.download, path: getDownloadPath() + "\\" + download.filename } ]);
-				currentDownloads.update((downloads) => {
-					const matchingDownload = downloads.find((d) => d.id === localID + '')!!;
-					matchingDownload.status = 'downloading';
-					matchingDownload.id = downloadID;
-					matchingDownload.usedRealDebrid = true;
-
-					matchingDownload.downloadPath = getDownloadPath() + "\\" + download.filename;
-					downloads[downloads.indexOf(matchingDownload)] = matchingDownload;
-					return downloads;
-				});
-				break;
-			}
-			case "real-debrid-torrent": {
-				if (!result.name || !result.downloadURL) {
-					createNotification({
-						id: Math.random().toString(36).substring(7),
-						type: 'error',
-						message: "Addon did not provide a name for the torrent."
-					});
-					return;
-				}
-
-				const worked = await window.electronAPI.realdebrid.updateKey();
-				if (!worked) {
-					notifications.update((notifications) => [...notifications, { id: Math.random().toString(36).substring(7), type: 'error', message: "Please set your Real-Debrid API key in the settings." }]);
-					return;
-				}
-				// add torrent link
-				const localID = Math.floor(Math.random() * 1000000);
-				currentDownloads.update((downloads) => {
-					return [...downloads, { 
-						id: '' + localID, 
-						status: 'rd-downloading', 
-						downloadPath: getDownloadPath() + "\\" + result.name, 
-						downloadSpeed: 0,
-						usedRealDebrid: true,
-						progress: 0,
-						...result 
-					}];
-				});
-				const torrent = await window.electronAPI.realdebrid.addTorrent(result.downloadURL);
-				const isReady = await window.electronAPI.realdebrid.isTorrentReady(torrent.id);
-				if (!isReady) {
-					window.electronAPI.realdebrid.selectTorrent(torrent.id);
-					await new Promise<void>((resolve) => {
-						const interval = setInterval(async () => {
-							const isReady = await window.electronAPI.realdebrid.isTorrentReady(torrent.id);
-							if (isReady) {
-								clearInterval(interval);
-								resolve();
-							}
-						}, 1000);
-					});
-				}
-
-
-				const torrentInfo = await window.electronAPI.realdebrid.getTorrentInfo(torrent.id);
-				// currently only supporting the first link
-				const download = await window.electronAPI.realdebrid.unrestrictLink(torrentInfo.links[0]);
-				if (download === null) {
-					createNotification({
-						id: Math.random().toString(36).substring(7),
-						type: 'error',
-						message: "Failed to unrestrict the link."
-					});
-					return;
-				}
-
-				const downloadID = await window.electronAPI.ddl.download([ { link: download.download, path: getDownloadPath() + "\\" + download.filename } ]);
-				currentDownloads.update((downloads) => {
-					const matchingDownload = downloads.find((d) => d.id === localID + '')!!
-					matchingDownload.status = 'downloading';
-					matchingDownload.id = downloadID;
-					matchingDownload.usedRealDebrid = true;
-
-					matchingDownload.downloadPath = getDownloadPath() + "\\" + download.filename;
-					downloads[downloads.indexOf(matchingDownload)] = matchingDownload;
-					return downloads;
-				});
-				break;
-			}
-			case 'torrent': {
-				if (!result.filename || !result.downloadURL) {
-						createNotification({
-							id: Math.random().toString(36).substring(7),
-							type: 'error',
-							message: "Addon did not provide a filename for the torrent."
-						});
-					return;
-				}
-				const downloadID = await window.electronAPI.torrent.downloadTorrent(result.downloadURL, getDownloadPath() + "\\" + (result.filename || result.downloadURL.split(/\\|\//).pop()));
-				if (downloadID === null) {
-					htmlButton.querySelector('[data-dwtext]')!!.textContent = "Download";
-					htmlButton.disabled = false;
-					return;
-				}
-				currentDownloads.update((downloads) => {
-					return [...downloads, { 
-						id: downloadID, 
-						status: 'downloading', 
-						downloadPath: getDownloadPath() + "\\" + result.filename + ".torrent", 
-						downloadSpeed: 0,
-						progress: 0,
-						usedRealDebrid: false,
-						...result 
-					}];
-				});
-				break;
-			}
-
-			case 'magnet': {
-				if (!result.filename || !result.downloadURL) {
-					createNotification({
-						id: Math.random().toString(36).substring(7),
-						type: 'error',
-						message: "Addon did not provide a filename for the magnet link."
-					});
-					return;
-				}
-
-				const downloadID = await window.electronAPI.torrent.downloadMagnet(result.downloadURL!!, getDownloadPath() + "\\" + result.filename!!);
-				if (downloadID === null) {
-					htmlButton.querySelector('[data-dwtext]')!!.textContent = "Download";
-					htmlButton.disabled = false;
-					return;
-				}
-				currentDownloads.update((downloads) => {
-					return [...downloads, { 
-						id: downloadID, 
-						status: 'downloading', 
-						downloadPath: getDownloadPath() + "\\" + result.filename + ".torrent", 
-						downloadSpeed: 0,
-						progress: 0,
-						usedRealDebrid: false,
-						...result 
-					}];
-				});
-				break;
-			}
-
-			case "direct": {
-				if (!result.filename && !result.files) {
-						createNotification({
-							id: Math.random().toString(36).substring(7),
-							type: 'error',
-							message: "Addon did not provide a filename for the direct download."
-						});
-					return;
-				}
-
-				let collectedFiles = [ { path: getDownloadPath() + "\\" + result.filename!!, link: result.downloadURL!! } ];
-				if (result.files) {
-					collectedFiles = result.files.map((file) => {
-						return { path: getDownloadPath() + "\\" + file.name, link: file.downloadURL };
-					});
-				}
-
-				const downloadID = await window.electronAPI.ddl.download(collectedFiles);
-				if (downloadID === null) {
-					htmlButton.querySelector('[data-dwtext]')!!.textContent = "Download";
-					htmlButton.disabled = false;
-					return;
-				}
-				currentDownloads.update((downloads) => {
-					return [...downloads, { 
-						id: downloadID, 
-						status: 'downloading', 
-						downloadPath: (result.files ? getDownloadPath() + "\\" : getDownloadPath() + "\\" + result.filename), 
-						downloadSpeed: 0,
-						usedRealDebrid: false,
-						progress: 0,
-						...result 
-					}];
-				});
-				break;
-			}
-		}
-	}
+	
 </script>
 <input id="search" on:change={search} type="text" placeholder="Search for Game" class="p-2 pl-2 bg-slate-100 rounded-lg w-2/3"/>
 {#if loadingResults}
